Added support for daily free pull banners.
Centralised ExecGacha and ExecuteEventGacha character pull logic into a utility class called GachaUtils to avoid repeating the same code multiple times.
