+++ EpinelPS/LobbyServer/Gacha/CheckGachaDailyEvent.cs
@@ -1,4 +1,8 @@
-﻿namespace EpinelPS.LobbyServer.Gacha;
+﻿using EpinelPS.Data;
+using EpinelPS.Utils;
+using Microsoft.EntityFrameworkCore.Metadata;
+
+namespace EpinelPS.LobbyServer.Gacha;
 
 [GameRequest("/gacha/event/check")]
 public class CheckGachaDailyEvent : LobbyMessage
@@ -6,21 +10,31 @@ public class CheckGachaDailyEvent : LobbyMessage
     protected override async Task HandleAsync()
     {
         ReqCheckDailyFreeGacha req = await ReadData<ReqCheckDailyFreeGacha>();
-        ResCheckDailyFreeGacha response = new()
+               
+
+        ResCheckDailyFreeGacha response = new ResCheckDailyFreeGacha();
+
+        var user = GetUser();
+
+        try
         {
-            // TODO implement
-            FreeCount = 1,
-            EventData = new NetEventData()
+            var evntFree = GameData.Instance.eventManagers[GameData.Instance.gachaTypes[req.GachaId].DailyFreeGachaEventId];
+
+            response.FreeCount = user.GachaDailyFreePulls.Contains(evntFree.Id) ? 0 : 1 ;
+            response.EventData = new NetEventData()
             {
-                Id = 80005,
-                EventSystemType = 21,
+                Id = evntFree.Id,
+                EventSystemType = (int)evntFree.EventSystemType,
                 EventVisibleDate = DateTime.UtcNow.Subtract(TimeSpan.FromDays(7)).Ticks,
                 EventStartDate = DateTime.UtcNow.Subtract(TimeSpan.FromDays(1)).Ticks,
                 EventEndDate = DateTime.Now.AddDays(20).Ticks,
                 EventDisableDate = DateTime.Now.AddDays(20).Ticks
-            }
-        };
-        // this is net event data i think it should be the same as in list events and get joined event only for free pull gacha event 
+            };
+        }
+        catch (Exception ex)
+        {
+            Logging.WriteLine("[CheckGachaDailyEvent] Event was not found in the database.");
+        }
 
         await WriteDataAsync(response);
     }
